move test logic into separate classes
