+++ release_tester/arangodb/starter/deployments/cluster.py
@@ -119,10 +119,7 @@ db.testCollection.save({test: "document"})
 
     def test_setup_impl(self):
         if self.selenium:
-            self.selenium.set_instances(self.cfg,
-                                        self.starter_instances[0].arango_importer,
-                                        self.starter_instances[0].arango_restore)
-            self.selenium.check_full_ui(self.passvoid, self.starter_instances)
+            self.selenium.test_setup()
 
     def wait_for_restore_impl(self, backup_starter):
         for starter in self.starter_instances:
@@ -222,7 +219,7 @@ db.testCollection.save({test: "document"})
 
         prompt_user(self.basecfg, "instance stopped")
         if self.selenium:
-            self.selenium.jam_step_1(self.new_cfg if self.new_cfg else self.cfg)
+            self.selenium.jam_step_1()
 
         ret = self.starter_instances[0].arangosh.check_test_data(
                 "Cluster one node missing", True, ['--disabledDbserverUUID', uuid] )
@@ -270,7 +267,8 @@ db.testCollection.save({test: "document"})
 
         prompt_user(self.basecfg, "cluster should be up")
         if self.selenium:
-            self.selenium.jam_step_2(self.new_cfg if self.new_cfg else self.cfg)
+            self.selenium.jam_step_2()
+            self.selenium.test_before_upgrade()
 
     def shutdown_impl(self):
         for node in self.starter_instances:
@@ -282,3 +280,12 @@ db.testCollection.save({test: "document"})
 
     def after_backup_impl(self):
         pass
+
+    def set_selenium_instances(self):
+        """ set instances in selenium runner """
+        self.selenium.set_instances(self.cfg,
+                                    self.starter_instances[0].arango_importer,
+                                    self.starter_instances[0].arango_restore,
+                                    [x for x in self.starter_instances[0].all_instances \
+                                     if x.instance_type == InstanceType.COORDINATOR][0],
+                                    self.new_cfg)
