+++ release_tester/arangodb/starter/deployments/selenium_deployments/activefailover.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python3
 """ test the UI of a active failover setup """
-import time
-import pprint
 from arangodb.starter.deployments.selenium_deployments.sbase import SeleniumRunner
-from release_tester.selenium_ui_test.pages.navbar import NavigationBarPage
-from reporting.reporting_utils import step
+from selenium_ui_test.test_suites.activefailover.before_upgrade_test_suite import \
+    ActiveFailoverBeforeUpgradeTestSuite
+from selenium_ui_test.test_suites.activefailover.jam_1_test_suite import \
+    ActiveFailoverJamStepOneSuite
+from selenium_ui_test.test_suites.basic_test_suite import BasicTestSuite
 
 
 class ActiveFailover(SeleniumRunner):
@@ -18,43 +19,8 @@ class ActiveFailover(SeleniumRunner):
                          is_headless,
                          testrun_name,
                          ssl)
-
-    @step
-    def check_old(self, cfg, leader_follower=False, expect_follower_count=2, retry_count=10):
-        """ check the integrity of the old system before the upgrade """
-        self.check_version(cfg)
-
-        while retry_count > 0:
-            NavigationBarPage(self.webdriver).navbar_goto('replication')
-            replication_table = self.get_replication_screen(True)
-            print(replication_table)
-            if len(replication_table['follower_table']) != expect_follower_count + 1:
-                time.sleep(5)
-                retry_count -= 1
-            else:
-                retry_count = 0 # its there!
-        # head and two followers should be there:
-        self.progress(' expecting %d followers, have %d followers'%(
-            expect_follower_count, len(replication_table['follower_table']) - 1))
-        self.ui_assert(len(replication_table['follower_table']) == expect_follower_count + 1,
-                       "UI-Test:\nexpect 1 follower in:\n%s" % pprint.pformat(
-                           replication_table))
-
-    @step
-    def upgrade_deployment(self, old_cfg, new_cfg, timeout):
-        """ nothing to see here """
-
-    @step
-    def jam_step_1(self, cfg, frontend_instance):
-        """ check for one set of instances to go away """
-        NavigationBarPage(self.webdriver).navbar_goto('replication')
-        replication_table = self.get_replication_screen(True)
-        print(replication_table)
-        # head and one follower should be there:
-        self.ui_assert(len(replication_table['follower_table']) == 2,
-                       "UI-Test:\nexpect 2 followers in:\n %s" % pprint.pformat(
-                           replication_table))
-        self.check_full_ui(cfg, frontend_instance)
-
-    def jam_step_2(self, cfg):
-        pass
+        self.main_test_suite_list = [BasicTestSuite]
+        self.before_upgrade_test_suite_list = [ActiveFailoverBeforeUpgradeTestSuite]
+        self.after_upgrade_test_suite_list = []
+        self.jam_step_1_test_suite_list = [ActiveFailoverJamStepOneSuite]
+        self.jam_step_2_test_suite_list = []
