+++ release_tester/arangodb/starter/deployments/selenium_deployments/leaderfollower.py
@@ -3,9 +3,10 @@
 import time
 from arangodb.starter.deployments.selenium_deployments.sbase import SeleniumRunner
 
-from reporting.reporting_utils import step
-
-from release_tester.selenium_ui_test.pages.navbar import NavigationBarPage
+from selenium_ui_test.test_suites.basic_test_suite import BasicTestSuite
+from selenium_ui_test.test_suites.leader_follower.before_upgrade_test_suite import \
+    LeaderFollowerBeforeUpgradeTestSuite
+from selenium_ui_test.test_suites.leader_follower.jam_1_test_suite import LeaderFollowerJamStepOneSuite
 
 
 class LeaderFollower(SeleniumRunner):
@@ -19,42 +20,10 @@ class LeaderFollower(SeleniumRunner):
                          is_headless,
                          testrun_name,
                          ssl)
-
-    @step
-    def check_old(self, cfg, leader_follower=True, expect_follower_count=2, retry_count=10):
-        """ check the integrity of the old system before the upgrade """
-        self.check_version(cfg)
-
-        count = 0
-        replication_table = None
-        while True:
-            NavigationBarPage(self.webdriver).navbar_goto('replication')
-            replication_table = self.get_replication_screen(leader_follower, 120)
-            self.progress(' ' + str(replication_table))
-            if len(replication_table['follower_table']) == 2:
-                break
-            if count % 5 == 0:
-                self.webdriver.refresh()
-            count +=1
-            time.sleep(5)
-        # head and one follower should be there:
-        self.ui_assert(len(replication_table['follower_table']) == 2,
-                       "UI-Test: expected 1 follower")
+        self.main_test_suite_list = [BasicTestSuite]
+        self.before_upgrade_test_suite_list = [LeaderFollowerBeforeUpgradeTestSuite]
+        self.jam_step_1_test_suite_list = [LeaderFollowerJamStepOneSuite]
 
     def upgrade_deployment(self, old_cfg, new_cfg, timeout):
         """ nothing to see here """
 
-    @step
-    def jam_step_1(self, cfg):
-        """ check for one set of instances to go away """
-        # TODO self.check_full_ui(cfg)
-        NavigationBarPage(self.webdriver).navbar_goto('replication')
-        replication_table = self.get_replication_screen(True)
-        print(replication_table)
-        # head and one follower should be there:
-        self.ui_assert(len(replication_table['follower_table']) == 2,
-                       "UI-Test: expected to have 1 follower!")
-        # TODO self.check_full_ui(cfg)
-
-    def jam_step_2(self, cfg):
-        """ nothing to see here """
