+++ release_tester/selenium_ui_test/pages/login_page.py
@@ -1,9 +1,6 @@
-#!/usr/bin/env python
-"""
-login to the UI
-"""
 import time
 
+from selenium.webdriver.support.select import Select
 from selenium_ui_test.pages.base_page import BasePage
 
 from selenium.webdriver.common.by import By
@@ -26,19 +23,16 @@ class LoginPage(BasePage):
         self.username_textbox_id = "loginUsername"
         self.password_textbox_id = "loginPassword"
         self.login_button_id = "submitLogin"
-        self.select_db_opt_id = "loginDatabase"
+        self.database_select = """//select[@id="loginDatabase"]"""
+        # self.select_db_opt_id = "loginDatabase"
         self.select_db_btn_id = "goToDatabase"
-
-    def progress(self, arg):
-        """ state print todo """
-        print(arg)
 
     def _login_wait_for_screen(self):
         """ wait for the browser to show the login screen """
         count = 0
         while True:
             count += 1
-            elem = WebDriverWait(self.driver, 10).until(
+            elem = WebDriverWait(self.webdriver, 10).until(
                 EC.presence_of_element_located((By.TAG_NAME, "html")),
                 message="UI-Test: page didn't load after 10s"
             )
@@ -50,45 +44,32 @@ class LoginPage(BasePage):
             if count == 10:
                 if elem is None:
                     self.progress(" locator has not been found.")
-                    self.driver.refresh()
+                    self.webdriver.refresh()
                     time.sleep(5)
                 else:
-                    assert "ArangoDB Web Interface" in self.driver.title, \
+                    assert "ArangoDB Web Interface" in self.webdriver.title, \
                         "webif title not found"
                     break
         return True
 
-    def _login_fill_username(self, cfg, user, passvoid, database, recurse=0):
+    def _login_fill_username(self, user):
         """ fill in the username column """
-        try:
-            logname = WebDriverWait(self.driver, 10).until(
-                EC.element_to_be_clickable((By.ID, "loginUsername")),
-                message="UI-Test: loginUsername didn't become clickeable on time. 10s"
-            )
-            logname.click()
-            logname.clear()
-            logname.send_keys(user)
-
-            if logname is None:
-                self.progress("locator loginUsername has not found.")
-                return False
-
-        except StaleElementReferenceException as ex:
-            self.progress("stale element, force reloading with sleep: " +
-                          str(ex))
-            self.driver.refresh()
-            time.sleep(5)
-            return self.login_webif(cfg,
-                                    user,
-                                    passvoid,
-                                    database, 
-                                    recurse + 1)
+        logname = WebDriverWait(self.webdriver, 10).until(
+            EC.element_to_be_clickable((By.ID, "loginUsername")),
+            message="UI-Test: loginUsername didn't become clickeable on time. 10s"
+        )
+        logname.click()
+        logname.clear()
+        logname.send_keys(user)
+        if logname is None:
+            self.progress("locator loginUsername has not found.")
+            return False
         return True
 
     def _login_fill_passvoid(self, passvoid):
         """ fill the passvoid and click login """
         while True:
-            passvoid_elm = self.driver.find_element_by_id("loginPassword")
+            passvoid_elm = self.webdriver.find_element_by_id("loginPassword")
             txt = passvoid_elm.text
             print("UI-Test: xxxx [" + txt + "]")
             if len(txt) > 0:
@@ -105,54 +86,28 @@ class LoginPage(BasePage):
             break
         return True
 
-    def _login_choose_database(self, cfg, user, passvoid, database, recurse=0):
+    def _login_choose_database(self, database_name):
         """ choose the database from the second login screen """
-        count = 0
-        while True:
-            count += 1
-            elem = WebDriverWait(self.driver, 15).until(
-                EC.presence_of_element_located((By.ID, "loginDatabase")),
-                message="UI-Test: loginDatabase didn't become clickeable on time 15s"
-            )
-            txt = elem.text
-            if txt.find(database) < 0:
-                self.progress(database + ' not found in ' +
-                              txt +
-                              ' ; retrying!')
-                if count == 10:
-                    self.progress('refreshing webpage and retrying...')
-                    self.driver.refresh()
-                    time.sleep(5)
-                    return self.login_webif(cfg,
-                                            user,
-                                            passvoid,
-                                            database,
-                                            recurse + 1)
-                time.sleep(2)
-            else:
-                break
-        elem = WebDriverWait(self.driver, 15).until(
-            EC.element_to_be_clickable((By.ID, "goToDatabase")),
-            message="UI-Test: choosing database didn't become clickeable on time 15s"
-        )
-        elem.click()
+        select = Select(self.locator_finder_by_xpath(self.database_select))
+        select.select_by_visible_text(database_name)
         return True
 
-    def login_webif(self, cfg, user, passvoid, database="_system", recurse=0):
+    def login_webif(self, user, passvoid, database="_system", recurse=0):
         """ log into an arangodb webinterface """
         print("Logging %s into %s with passvoid %s" %(user, database, passvoid))
         if recurse > 10:
             raise Exception("UI-Test: 10 successless login attempts")
         self._login_wait_for_screen()
-        if not self._login_fill_username(cfg, user, passvoid, database, recurse):
+        if not self._login_fill_username(user):
             return False
         if not self._login_fill_passvoid(passvoid):
             return False
-        if not self._login_choose_database(cfg, user, passvoid, database, recurse):
+        if not self._login_choose_database(database):
             return False
+        self.webdriver.find_element_by_id(self.select_db_btn_id).click()
         self.progress("we're in!")
         self.wait_for_ajax()
 
-        assert "No results found." not in self.driver.page_source, \
+        assert "No results found." not in self.webdriver.page_source, \
             "no results found?"
         return False
